+++ lib/collection.dart
@@ -1,9 +1,33 @@
-/// A collection of new collection types: bi-map, bit-list, multi-set, set and
-/// list multi-map, range, and string.
+/// A collection of iterable extensions and new collection types: bi-map,
+/// bit-list, multi-set, set and list multi-map, range, and string.
 export 'src/collection/bimap.dart'
     show BiMap, BiMapOnMapExtension, BiMapOnIterableExtension;
 export 'src/collection/bitlist.dart' show BitList, BitListExtension;
 export 'src/collection/heap.dart' show Heap;
+export 'src/collection/iterable/chunked.dart' show ChunkedExtension;
+export 'src/collection/iterable/combinations.dart' show CombinationsExtension;
+export 'src/collection/iterable/count.dart' show CountIterableExtension;
+export 'src/collection/iterable/flat_map.dart' show FlatMapExtension;
+export 'src/collection/iterable/flatten.dart' show FlattenExtension;
+export 'src/collection/iterable/group.dart' show GroupExtension, Group;
+export 'src/collection/iterable/indexed.dart' show IndexedExtension, Indexed;
+export 'src/collection/iterable/iterate.dart' show iterate;
+export 'src/collection/iterable/operators.dart' show OperatorsExtension;
+export 'src/collection/iterable/periodical.dart'
+    show Period, PeriodicalExtension, TruncateExtension;
+export 'src/collection/iterable/permutations.dart' show PermutationsExtension;
+export 'src/collection/iterable/product.dart' show ProductExtension;
+export 'src/collection/iterable/random.dart' show RandomExtension;
+export 'src/collection/iterable/repeat_element.dart' show repeat;
+export 'src/collection/iterable/repeat_iterable.dart'
+    show RepeatIterableExtension;
+export 'src/collection/iterable/separated.dart' show SeparatedExtension;
+export 'src/collection/iterable/to_map.dart' show ToMapExtension;
+export 'src/collection/iterable/unique.dart' show UniqueExtension;
+export 'src/collection/iterable/window.dart' show WindowExtension;
+export 'src/collection/iterable/zip.dart' show ZipExtension;
+export 'src/collection/list/rotate.dart'
+    show RotateListExtension, RotateQueueExtension;
 export 'src/collection/list/take_skip.dart' show TakeSkipListExtension;
 export 'src/collection/multimap/list.dart'
     show