+++ test/all_test.dart
@@ -9,7 +9,6 @@ import 'feature_test.dart' as feature_test;
 import 'functional_test.dart' as functional_test;
 import 'functional_type_test.dart' as functional_type_test;
 import 'interval_test.dart' as interval_test;
-import 'iterable_test.dart' as iterable_test;
 import 'math_test.dart' as math_test;
 import 'number_test.dart' as number_test;
 import 'ordering_test.dart' as ordering_test;
@@ -27,7 +26,6 @@ void main() {
   group('functional', functional_test.main);
   group('functional type', functional_type_test.main);
   group('interval', interval_test.main);
-  group('iterable', iterable_test.main);
   group('math', math_test.main);
   group('number', number_test.main);
   group('ordering', ordering_test.main);